set tooltip helper methods in a tooltip_helper
