+++ app/helpers/application_helper.rb
@@ -51,32 +51,6 @@ module ApplicationHelper
     Rack::Utils.parse_nested_query(request.query_string).deep_symbolize_keys
   end
 
-  def tooltip_tag_attributes(stimulus_action:, **dataset)
-    attributes = {
-      data: {
-        controller: "tooltip",
-        action: stimulus_action
-      }.merge(dataset || {})
-    }
-
-    tag.attributes(attributes)
-  end
-
-  def tooltip_errors_tag_attributes(title:, errors:)
-    tag.attributes(tooltip_errors_attributes(title: title, errors: errors))
-  end
-
-  def tooltip_errors_attributes(title:, errors:)
-    {
-      data: {
-        controller: "tooltip",
-        action: "mouseover->tooltip#showErrors",
-        title: title,
-        errors: errors.to_json
-      }
-    }
-  end
-
   def render_turbo_stream_flash_messages
     turbo_stream.prepend "flashes", partial: "common/flashes"
   end
